Update UI: luxury redesign, new homepage sections, hero/testimonial sliders, admin fixes
- Full luxury visual redesign of the storefront (grey/black/white/pink theme):
  3D coverflow, animated ambient background, floating WhatsApp/custom-bouquet
  actions, real image sliders for hero and About Us sections
- New homepage sections: Why Choose Us, About Us teaser, What Our Customers
  Say (backed by a new public GET /reviews/featured endpoint)
- Generic infinite carousel component replacing the old hero/product carousels,
  now auto-playing on all screen sizes with hover-to-pause
- Smoother hover/tilt interactions across cards and images
- Fixed a real bug where creating a product via the admin panel always failed
  (stock/lowStockThreshold/weightGrams/quantity weren't coerced from
  form-data strings to Prisma Int fields)
- Fixed position:sticky header breaking due to body overflow-x:hidden
- Added gzip/brotli compression to the Angular SSR server, lazy-loading on
  below-the-fold images, and a realistic production style-budget
- Various asset/branding fixes (logo, favicon)

+++ src/modules/products/products.validation.js
@@ -27,9 +27,9 @@ const createRules = [
   body('basePrice').isFloat({ min: 0 }).withMessage('Base price must be a positive number'),
   body('discountPrice').optional({ nullable: true }).isFloat({ min: 0 }),
   body('costPrice').optional({ nullable: true }).isFloat({ min: 0 }),
-  body('stock').optional().isInt({ min: 0 }),
-  body('lowStockThreshold').optional().isInt({ min: 0 }),
-  body('weightGrams').optional().isInt({ min: 0 }),
+  body('stock').optional().isInt({ min: 0 }).toInt(),
+  body('lowStockThreshold').optional().isInt({ min: 0 }).toInt(),
+  body('weightGrams').optional().isInt({ min: 0 }).toInt(),
   body('isActive').optional().isBoolean().toBoolean(),
   body('isFeatured').optional().isBoolean().toBoolean(),
   body('shortDescription').optional().trim().isLength({ max: 500 }),
@@ -46,9 +46,9 @@ const updateRules = [
   body('basePrice').optional().isFloat({ min: 0 }),
   body('discountPrice').optional({ nullable: true }).isFloat({ min: 0 }),
   body('costPrice').optional({ nullable: true }).isFloat({ min: 0 }),
-  body('stock').optional().isInt({ min: 0 }),
-  body('lowStockThreshold').optional().isInt({ min: 0 }),
-  body('weightGrams').optional().isInt({ min: 0 }),
+  body('stock').optional().isInt({ min: 0 }).toInt(),
+  body('lowStockThreshold').optional().isInt({ min: 0 }).toInt(),
+  body('weightGrams').optional().isInt({ min: 0 }).toInt(),
   body('isActive').optional().isBoolean().toBoolean(),
   body('isFeatured').optional().isBoolean().toBoolean(),
   body('shortDescription').optional().trim().isLength({ max: 500 }),
@@ -59,7 +59,7 @@ const variantRules = [
   body('name').trim().notEmpty().withMessage('Variant name is required'),
   body('sku').trim().notEmpty().withMessage('Variant SKU is required'),
   body('priceAdjustment').optional().isFloat(),
-  body('stock').optional().isInt({ min: 0 }),
+  body('stock').optional().isInt({ min: 0 }).toInt(),
   body('isDefault').optional().isBoolean().toBoolean(),
 ];
 
@@ -67,14 +67,14 @@ const variantUpdateRules = [
   body('name').optional().trim().notEmpty(),
   body('sku').optional().trim().notEmpty(),
   body('priceAdjustment').optional().isFloat(),
-  body('stock').optional().isInt({ min: 0 }),
+  body('stock').optional().isInt({ min: 0 }).toInt(),
   body('isDefault').optional().isBoolean().toBoolean(),
   body('isActive').optional().isBoolean().toBoolean(),
 ];
 
 const stockAdjustRules = [
   body('mode').isIn(['SET', 'INCREMENT', 'DECREMENT']).withMessage('mode must be SET, INCREMENT or DECREMENT'),
-  body('quantity').isInt({ min: 0 }).withMessage('quantity must be a non-negative integer'),
+  body('quantity').isInt({ min: 0 }).withMessage('quantity must be a non-negative integer').toInt(),
 ];
 
 module.exports = {
